requirements.txt
anthropic>=0.35.0
feedparser>=6.0.0
requests>=2.31.0
pyyaml>=6.0.0
jinja2>=3.1.0
